Remove Collection Methods
Refactored the Solutions to no longer extend collection and be more abstract.

This is useful for more and different kinds of solutions they do not need to be tied to collection implementation and instead can be interfaced solely through place codes

+++ src/Test/Main.java
@@ -18,7 +18,7 @@ import staticMethods.SolutionMatcher;
 public class Main {
 
 	public static void main(String[] args) throws InterruptedException {
-		NumericAlgorithm<Double> na = new NumericAlgorithm<Double>(1, new CircleProblem<Double>(), 3);
+		NumericAlgorithm<Integer> na = new NumericAlgorithm<Integer>(1, new CircleProblem<Integer>(), 3);
 		na.setAlgType(2);
 		na.generateSolutions(10);
 		na.setParticleSwarmOpt(1, 0.25, 0.25, 0.25, 0.25);
